requirements.txt
pyyaml
flask
aiohttp
discord